src/fraud_detection/requirements.txt
scikit-learn==1.4.1.post1
pandas==2.2.1
polars==0.20.15
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
pyarrow==15.0.1
setuptools==69.2.0
lightgbm==4.3.0